runstatus: split things up
